Moving GroupBox accessible object to a separate file.
